rename: ToolTip 삼각형 asset 디렉토리 위치 이동
